src/backend/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
azure-storage-blob[aio]==12.19.1
azure-identity==1.15.0
azure-search-documents==11.6.0b3
azure-ai-formrecognizer==3.3.3
pypdf==3.17.0
python-dotenv==1.0.1
opentelemetry-instrumentation-flask==0.40b0
azure-monitor-opentelemetry==1.0.0b15
openai==1.25.0
pdf2image==1.17.0
pytz==2024.1
docx2pdf==0.1.8
pywin32==306